+++ lib/flutter_network_connectivity.dart
@@ -11,6 +11,8 @@ import 'package:flutter/services.dart';
 class FlutterNetworkConnectivity {
   static const _defaultLookUpUrl = 'example.com';
 
+  final Duration _defaultDuration = const Duration(seconds: 5);
+
   /// Method Channel to Check if Network is Available
   /// To check one time call
   static const MethodChannel _methodChannel = MethodChannel(
@@ -23,18 +25,49 @@ class FlutterNetworkConnectivity {
   /// URL to lookup for to check network connection availability
   late String _lookUpUrl;
 
+  /// set to true in constructor to lookup in a regular interval
+  late bool _isContinousLookUp;
+
+  /// override default lookup duration in constructor params
+  late Duration _lookUpDuration;
+
   /// check if internet is available using this bool
   bool isInternetAvailable = false;
 
-  /// add custom url [lookUpUrl] to lookup instead of [_defaultLookUpUrl]
-  FlutterNetworkConnectivity({String? lookUpUrl}) {
-    _lookUpUrl = lookUpUrl ?? _defaultLookUpUrl;
+  /// stream resulting true/false based on connectivity changes,
+  /// lookup on interval if [_isContinousLookUp] is set true
+  late StreamController<bool> _internetAvailabilityStreamController;
 
-    _init();
+  /// timer to handle continous look up if [_isContinousLookUp] is set true
+  Timer? _timer;
+
+  FlutterNetworkConnectivity(
+      {String? lookUpUrl,
+      bool isContinousLookUp = false,
+      Duration? lookUpDuration}) {
+    _internetAvailabilityStreamController = StreamController<bool>();
+
+    _lookUpUrl = lookUpUrl ?? _defaultLookUpUrl;
+    _isContinousLookUp = isContinousLookUp;
+    _lookUpDuration = lookUpDuration ?? _defaultDuration;
+
+    if (_isContinousLookUp) {
+      _startContinousLookUp();
+    }
+  }
+
+  /// Get Internet Availability Stream based on config's
+  /// Default Config: Get Stream on Connectivity Status Change
+  /// Available Config's: (to be configured in constructor while object creation)
+  /// 1. Set Internet Availability on Period Check [isContinousLookUp] in constructor, default is false
+  /// 2. Duration of the Period Check, [lookUpDuration] from constructor, default is [_defaultDuration]
+  /// 3. Custom Url to Look Up, [lookUpUrl] from constructor, default is [_defaultLookUpUrl]
+  Stream<bool> getInternetAvailabilityStream() {
+    return _internetAvailabilityStreamController.stream;
   }
 
   /// Registers Network Connectivity Listener from Native
-  void _init() async {
+  Future registerAvailabilityListener() async {
     /// Listen for Network Connectivity Status
     _getNetworkConnectivityStatusStream().listen((isNetworkConnected) {
       _checkConnection();
@@ -44,7 +77,10 @@ class FlutterNetworkConnectivity {
     await _checkConnection();
   }
 
-  Future unregisterListener() async {
+  /// Unregister to cancel timer if set,
+  /// close connection listeners from native calls
+  Future unregisterAvailabilityListener() async {
+    _timer?.cancel();
     await _unregisterNetworkConnectivityListener();
   }
 
@@ -52,7 +88,21 @@ class FlutterNetworkConnectivity {
   ///
   /// Android call requires permission [android.permission.ACCESS_NETWORK_STATE]
   /// iOS min version required: 12.0
+  /// deprecation due to renaming of method, replacement method
+  /// [isNetworkConnectionAvailable]
+  @Deprecated(
+      'This Method will be deprecated on the next release, use isNetworkConnectionAvailable instead')
   Future<bool> isNetworkAvailable() async {
+    bool hasConnection =
+        await _methodChannel.invokeMethod("isNetworkAvailable");
+    return hasConnection;
+  }
+
+  /// Checks if Network is Available
+  ///
+  /// Android call requires permission [android.permission.ACCESS_NETWORK_STATE]
+  /// iOS min version required: 12.0
+  Future<bool> isNetworkConnectionAvailable() async {
     bool hasConnection =
         await _methodChannel.invokeMethod("isNetworkAvailable");
     return hasConnection;
@@ -62,7 +112,7 @@ class FlutterNetworkConnectivity {
   /// Must be unregistered if not in use
   /// registering only on android as iOS makes use of NetworkMonitor
   @Deprecated(
-      'This Method will be deprecated on the next release, will be registered internally')
+      'This Method will be deprecated on the next release, use registerConnectivityListener instead')
   Future registerNetworkListener() async {
     if (Platform.isAndroid) {
       await _methodChannel.invokeMethod("registerNetworkStatusListener");
@@ -72,19 +122,23 @@ class FlutterNetworkConnectivity {
   /// Unregister Listener when not in use
   /// unregister only on android as iOS makes use of NetworkMonitor
   @Deprecated(
-      'This method will be deprecated on the next release, will be unregistered internally')
+      'This method will be deprecated on the next release, use unregisterListener instead')
   Future unregisterNetworkListener() async {
     if (Platform.isAndroid) {
       await _methodChannel.invokeMethod("unregisterNetworkStatusListener");
     }
   }
 
+  /// Internally register connectivity change listener
+  /// only on android
   Future _registerNetworkConnectivityListener() async {
     if (Platform.isAndroid) {
       await _methodChannel.invokeMethod("registerNetworkStatusListener");
     }
   }
 
+  /// Internally unregister connectivity change listener
+  /// only on android
   Future _unregisterNetworkConnectivityListener() async {
     if (Platform.isAndroid) {
       await _methodChannel.invokeMethod("unregisterNetworkStatusListener");
@@ -100,26 +154,47 @@ class FlutterNetworkConnectivity {
     return _eventChannel.receiveBroadcastStream().cast<bool>();
   }
 
+  /// Platform connectivity status stream from native
   Stream<bool> _getNetworkConnectivityStatusStream() {
     return _eventChannel.receiveBroadcastStream().cast<bool>();
   }
 
+  /// Helper to for connection check
   Future _checkConnection() async {
     isInternetAvailable = await isInternetConnectionAvailable();
+    _updateStreamController(isInternetAvailable);
   }
 
+  /// use this method to check to check for internet connection availability
+  /// also used on internal methods on periodic/connection status changed
+  /// override [_lookUpUrl] on constructor while object creation
   Future<bool> isInternetConnectionAvailable() async {
     try {
       final result = await InternetAddress.lookup(_lookUpUrl);
       if (result.isNotEmpty && result[0].rawAddress.isNotEmpty) {
         isInternetAvailable = true;
+        _updateStreamController(true);
       } else {
         isInternetAvailable = false;
+        _updateStreamController(false);
       }
     } on SocketException catch (_) {
       isInternetAvailable = false;
+      _updateStreamController(false);
     }
 
     return isInternetAvailable;
+  }
+
+  /// Helper to update internet status to stream
+  _updateStreamController(bool value) {
+    _internetAvailabilityStreamController.add(value);
+  }
+
+  /// Starts period check if [_isContinousLookUp] is true from constructor
+  void _startContinousLookUp() {
+    _timer = Timer.periodic(_lookUpDuration, (timer) {
+      isInternetConnectionAvailable();
+    });
   }
 }